requirements.txt
# Core Dependencies
numpy==1.26.4
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78

# AirSim Communication
msgpack-rpc-python==0.4.1
msgpack==1.0.5
tornado==4.5.3
airsim==1.8.1
Pillow

# Neural Network and Perception (CPU version for AMD GPU)
# Using CPU version to avoid compatibility issues with torch-directml
torch==2.2.2
torchvision==0.17.2

# Additional Dependencies
tqdm
scipy
matplotlib
scikit-learn
